Fixed Account#getAccountNumber() method (it should return an integer)

+++ Laborator 6/BankAccounts/src/me/stefan923/bankaccounts/Account.java
@@ -36,7 +36,7 @@ public class Account {
         return balance;
     }
 
-    public double getAccountNumber() {
+    public int getAccountNumber() {
         return number;
     }
